Fix imports after moving AppBtn and AppIcon components
